ExcoVerse/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-bootstrap-datepicker-plus==5.0.4
django-datetime-widget==0.9.3
django-import-export==3.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pydantic==2.0.3
pydantic_core==2.3.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
xlrd==2.0.1
xlwt==1.3.0
stripe==5.5.0
aiogram==2.25.1
django-extensions==3.2.3